% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{acpc}
\alias{acpc}
\title{Area Classification on Dimension Reduced Micro-level Survey Data}
\usage{
acpc(
  X,
  K,
  r = 2,
  eps_r = 1e-04,
  MaxIter_r = 1e+05,
  gamma = 0.1,
  tau = 1,
  eps_s = 1e-04,
  MaxIter_s = 1000,
  lambda = 0.1,
  MaxIter_k = 1000
)
}
\arguments{
\item{X}{n * (p + 1) data frame contains first column for observation name ID and the rest p numeric features for n observations, column names are feature IDs.}

\item{K}{Scalar, Number of clusters.}

\item{r}{Scalar, number of principle components you want to generate, r<< p.}

\item{eps_r}{Scalar, tolerance for robust PCA.}

\item{MaxIter_r}{Scalar, maximum iteration for robust PCA.}

\item{gamma}{Scalar, relative weights between nuclear norm and sparsity penalties for robust PCA.}

\item{tau}{Scalar, scale parameter for ADMM updates in robust PCA.}

\item{eps_s}{Scalar, tolerance for sparse PCA.}

\item{MaxIter_s}{Scalar, maximum iteration for sparse PCA.}

\item{lambda}{Scalar, sparsity penalty parameter for sparse PCA.}

\item{MaxIter_k}{Scalar, maximum iteration for Kmeans.}
}
\value{
Label for clustering, Y; scores for observations, U; loadings for principle components, V.
}
\description{
Area Classification on Dimension Reduced Micro-level Survey Data
}
\examples{
# Create matrix M according to M = low rank + few large elements
n = 100
p = 40
# Create rank 3 component
set.seed(1234)
out <- svd(matrix(rnorm(n*p), n, p), nu = 3, nv = 3)
trueL <- out$u \%*\% diag(out$d[1:3]) \%*\% t(out$v)
# Create sparse component
trueS <- matrix(rt(n*p, df = 1), n, p)
trueS[abs(trueS) < 2.5] <- 0
sum(trueS !=0)
# Create M by combining the above
M <- trueL + trueS
# Verify that M itself is not small rank and not even close to being rank 3
svd(M)$d
K = 5
M <- cbind(data.frame(as.character(1:n), M))
acpc(M, K)
}
