% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{Title}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{Data matrix with dimensions n by p}

\item{K}{Number of clusters  with K less than or equal to n (optional)}

\item{M}{Matrix of initial cluster centers with dimensions K by p (optional)}

\item{numIter}{Number of iterations of algorithm (default = 100)}
}
\value{
Returns vector Y of length n, which assigns each data point into a cluster
}
\description{
Title
}
\examples{
# Example on 2D-plane
# Generate features.
X = rbind(matrix(rnorm(200, sd = 1), ncol = 2), matrix(rnorm(200, mean = 6, sd = 1), ncol = 2))
colnames(X) = c("x", "y")
# Cluster observations into 2 groups.
clusters = MyKmeans(X, 2)
plot(X, col = clusters)
}
