% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{runExample}
\alias{runExample}
\title{Interface for Area classification based on survey data.}
\usage{
runExample()
}
\value{
Run shiny app for output.
}
\description{
Interface for Area classification based on survey data.
}
