% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{acpc}
\alias{acpc}
\title{Title}
\usage{
acpc(X, Y = NULL, K = NULL, r, lambda = 0, eps = 1e-04, numIter = 100)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{K}{}

\item{r}{}

\item{lambda}{}

\item{eps}{}

\item{numIter}{}
}
\value{

}
\description{
Title
}
