% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Interface for Area classification based on census data.}
\usage{
visualization()
}
\value{
Run shiny app for output, two principle components, all parameters are set by default value.
}
\description{
You're supposed to save your data input as an object call "data_for_acpc" in your global environment, then do visualization() to see the App window.
}
