% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Interface for Area classification based on census data.}
\usage{
visualization()
}
\value{
Run shiny app for output, two principle components, all parameters are set by default value.
}
\description{
Interface for Area classification based on census data.
}
