% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualization}
\alias{visualization}
\title{Interface for Area classification based on survey data.}
\usage{
visualization(censusdata)
}
\arguments{
\item{censusdata}{Input data set for visualization.}
}
\value{
Run shiny app for output.
}
\description{
Interface for Area classification based on survey data.
}
